refactoring serve directory
