+++ src/serve/serveAction.ts
@@ -2,7 +2,7 @@ import { spawn } from "child_process";
 import { LiffApiClient } from "../api/liff.js";
 import { resolveChannel } from "../channel/resolveChannel.js";
 import { getCurrentChannelId } from "../channel/stores/channels.js";
-import { LocalProxy } from "../proxy/local-proxy.js";
+import { LocalProxy } from "./proxy/local-proxy.js";
 import resolveEndpointUrl from "./resolveEndpointUrl.js";
 import pc from "picocolors";
