+++ src/setup.ts
@@ -1,12 +1,12 @@
 import { Command } from "commander";
 import { installChannelCommands } from "./channel/commands/index.js";
 import { installAppCommands } from "./app/commands/index.js";
-import { serveCommands } from "./serve/index.js";
+import { installServeCommands } from "./serve/commands/index.js";
 
 export const setupCLI = (program: Command) => {
   installChannelCommands(program);
   installAppCommands(program);
-  serveCommands(program);
+  installServeCommands(program);
   // TODO .version?
   return {
     run: (argv = process.argv) => {
